sorting

+++ sort.h
@@ -26,7 +26,7 @@ void selection_sort(int *array, size_t size);
 void bubble_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
 void qck_sort(int *array, int l, int h, size_t size);
-int lomuto_partition(int *array, int l, int h, size_t size);
+int lomuto_partition(int *array, int l, int h, int size);
 void swap(int *a, int *b);
 void insertion_sort_list(listint_t **list);
 listint_t *swap_node(listint_t *nd, listint_t **list);
